cleanup loader logic, works in dockerfile now
