+++ sgrep_lint/sgrep.py
@@ -7,9 +7,25 @@ from constants import DEFAULT_CONFIG_FILE
 from constants import PLEASE_FILE_ISSUE_TEXT
 from constants import RCE_RULE_FLAG
 from constants import SGREP_URL
-from util import print_error_exit
+import config_resolver
+from util import print_error_exit, debug_print
+
+
+def set_flags(debug: bool, quiet: bool) -> None:
+    """Set the global DEBUG and QUIET flags"""
+    # TODO move to a proper logging framework
+    global DEBUG
+    global QUIET
+    if debug:
+        DEBUG = True  # type: ignore
+        debug_print("DEBUG is on")
+    if quiet:
+        QUIET = True  # type: ignore
+        debug_print("QUIET is on")
+
 
 # CLI
+
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(
@@ -130,6 +146,14 @@ if __name__ == "__main__":
     args = parser.parse_args()
     if args.lang and not args.pattern or (args.pattern and not args.lang):
         parser.error("-e/--pattern and -l/--lang must both be specified")
+
+
+    # set the flags
+    set_flags(args.verbose, args.quiet)
+
+    # change cwd if using docker
+    config_resolver.adjust_for_docker(args.precommit)
+
     try:
         if args.test:
             test.test_main(args)
